Move the parsing functions into a text module.
Move parse_text function as well as the parse_alias function into a
separate public module.
Also add & update some docs.
