+++ examples/parse_text.rs
@@ -1,16 +1,14 @@
-use emojic;
-
 const TEXT: &str = "
 Hello :waving_hand:, I am a :technologist:
-This :memo: is writen on a :desktop_computer: not with a :pencil:
+This :memo: is written on a :desktop_computer: not with a :pencil:
 This :memo: of course is not on a :floppy_disk: but on :computer_disk:
-We don't use :file_cabinet: but :file_folder: which can be :open_file_folder:
+Stored in a :file_folder: that can be :open_file_folder: instead of a :file_cabinet:
 ";
 
 fn main() {
     #[cfg(feature = "alloc")]
-    println!("{}", emojic::parse(TEXT));
+    println!("{}", emojic::text::parse_text(TEXT));
 
     #[cfg(not(feature = "alloc"))]
-    println!("{}", emojic::EmojiTextParser::new(TEXT));
+    println!("{}", emojic::text::EmojiTextParser::new(TEXT));
 }
