+++ src/text.rs
@@ -1,5 +1,134 @@
+//! Text processing utilities
+//!
+//! This module contains some utilities to enrich text with emojis.
+//! The basis of this module is the [`parse_alias`] function which takes
+//! emoji names (e.g. `:crab:`) and tries to look up the appropriate emoji
+//! (e.g. `🦀`).
+//!
+//! Based on the [`parse_alias`] function, the [`parse_text`] function takes
+//! an entire text and looks for such colon-fenced emoji names, which will then
+//! be translated into the Unicode equivalent, and the entire text, with the
+//! emojis replaced is returned.
+
 use core::fmt;
 
+use crate::emojis::Emoji;
+
+#[cfg(feature = "alloc")]
+use alloc::string::String;
+
+/// Parses the given Emoji name into a unicode Emoji.
+///
+/// This function accepts strings of the form `:name:` and looks up an emojis for it.
+/// The list of valid names is taken from: [github/gemoji](https://github.com/github/gemoji)
+/// And additonally all the constant names (as listed in [`crate::flat`]) are also valid aliases
+/// when spelled in lowercase.
+///
+/// # Examples
+///
+/// ```
+/// use emojic::parse_alias;
+///
+/// // gemoji style
+/// assert_eq!(
+///     Some(&*emojic::flat::THUMBS_UP),
+///     parse_alias(":+1:") //👍
+/// );
+///
+/// // constant name style
+/// assert_eq!(
+///     Some(&emojic::flat::ALIEN_MONSTER),
+///     parse_alias(":alien_monster:") //👾
+/// );
+/// ```
+///
+pub fn parse_alias(inp: &str) -> Option<&'static Emoji> {
+    // make some basic checks
+    if inp.starts_with(':') && inp.ends_with(':') && inp.is_ascii() && inp.len() > 2 {
+        // go on with the middle part
+        parse_pure_alias(&inp[1..(inp.len() - 1)])
+    } else {
+        None
+    }
+}
+
+/// Parses a pice of string into an emoji (no colons)
+fn parse_pure_alias(inp: &str) -> Option<&'static Emoji> {
+    cfg_if::cfg_if! {
+        if #[cfg(feature = "alloc")] {
+            // If we have alloc, we use the faster hash map
+            crate::alias::GEMOJI_MAP.get(inp).cloned()
+        } else {
+            // As a fallback, we can also use a huge match statement
+            crate::matching::matching(inp)
+        }
+    }
+}
+
+/// Replaces all gemojis (`:[a-z0-9_+-]+:`) found in `text` with their Unicode equivalent.
+///
+/// This function is a convenience function for [`EmojiTextParser`]:
+/// ```rust
+/// # use emojic::text::{EmojiTextParser,parse_text};
+/// # let text = ":some: :cat:";
+/// # assert_eq!(
+/// EmojiTextParser::new(text).to_string()
+/// # , parse_text(text)
+/// # );
+/// ```
+///
+/// Notice, this convenience function requires `alloc` unlike the
+/// [`EmojiTextParser`] iterator.
+///
+/// # Example
+///
+/// ```rust
+/// use emojic::text::parse_text;
+/// assert_eq!(
+///     &parse_text("Hello :waving_hand:, I am a :technologist:."),
+///     "Hello 👋, I am a 🧑‍💻.",
+/// );
+/// ```
+///
+/// ```rust
+/// use emojic::text::parse_text;
+/// assert_eq!(
+///     &parse_text("Neither std::iter::Iterator nor :rustaceans: are emojis"),
+///     "Neither std::iter::Iterator nor :rustaceans: are emojis",
+/// );
+/// ```
+#[cfg(feature = "alloc")]
+#[cfg_attr(feature = "doc_cfg", doc(cfg(feature = "alloc")))]
+pub fn parse_text(text: &str) -> String {
+    EmojiTextParser::new(text).collect()
+}
+
+/// Finds and replaces gemojis (`:[a-z0-9_+-]+:`) in text.
+///
+/// This is the iterator behind [`parse_text`].
+///
+/// `EmojiTextParser` simply splits its input text into multiple fragments,
+/// those of plain text, and those which are replacement emoji sequences.
+///
+/// Notice, that since this is simple iterator, it dose not depend on `alloc`,
+/// unlike the convenience function [`parse_text`].
+///
+/// # Example
+///
+/// ```rust
+/// use emojic::text::EmojiTextParser;
+///
+/// let input = "Hello :waving_hand:, I am a :technologist:.";
+/// let mut parser = EmojiTextParser::new(input);
+///
+/// assert_eq!(Some("Hello "), parser.next());
+/// assert_eq!(Some("👋"), parser.next());
+/// assert_eq!(Some(", I am a "), parser.next());
+/// assert_eq!(Some("🧑‍💻"), parser.next());
+/// assert_eq!(Some("."), parser.next());
+/// assert_eq!(None, parser.next());
+/// ```
+///
 #[derive(Debug, Clone)]
 pub struct EmojiTextParser<'a> {
     /// The original string that is parsed, used to fetch the output strings of
@@ -11,6 +140,7 @@ pub struct EmojiTextParser<'a> {
     emoji_fragment_start: bool,
 }
 impl<'a> EmojiTextParser<'a> {
+    /// Creates a new parser for the given `original` text.
     pub fn new(original: &'a str) -> Self {
         Self {
             original,
@@ -59,7 +189,6 @@ impl<'a> Iterator for EmojiTextParser<'a> {
                     let current_pos = self.next_pos + i;
                     // This is the second colon
 
-                    // TODO check bounds!
                     let emoji_name = &self.original[start_idx..=current_pos];
 
                     if let Some(e) = crate::parse_alias(emoji_name) {
@@ -222,5 +351,18 @@ mod tests {
         let parser = EmojiTextParser::new(input);
 
         assert_eq!(output, &parser.collect::<String>());
+    }
+
+    #[test]
+    fn parse_alias_test() {
+        assert_eq!(
+            Some(&crate::flat::FLAG_ECUADOR),
+            parse_alias(":flag_ecuador:")
+        );
+    }
+
+    #[test]
+    fn parse_alias_none() {
+        assert_eq!(None, parse_alias(":hebele:"));
     }
 }
